[PREVIEW] Vendor XLA/TSL into TF as bazel dependencies

+++ tensorflow/core/kernels/mkl/mkl_layer_norm_op.cc
@@ -15,7 +15,7 @@ limitations under the License.
 
 #ifdef INTEL_MKL
 
-#include "third_party/eigen3/unsupported/Eigen/CXX11/Tensor"
+#include "unsupported/Eigen/CXX11/Tensor"  // from @eigen_archive
 #include "dnnl.hpp"
 #include "tensorflow/core/framework/op_kernel.h"
 #include "tensorflow/core/framework/register_types.h"
@@ -53,8 +53,8 @@ class MklLayerNormOp : public OpKernel {
       OP_REQUIRES(ctx, shift_tensor.dims() == 1,
                   errors::InvalidArgument("offset must be 1D tensor",
                                           shift_tensor.shape().DebugString()));
-      int64_t num_elements_scale = scale_tensor.dim_size(0);
-      int64_t num_elements_shift = shift_tensor.dim_size(0);
+      size_t num_elements_scale = scale_tensor.dim_size(0);
+      size_t num_elements_shift = shift_tensor.dim_size(0);
       OP_REQUIRES(
           ctx, num_elements_scale == num_elements_shift,
           errors::InvalidArgument("Number of elements in scale and shift",
